requirements.txt
directory_tree
google-generativeai
python-dotenv
pdfminer.six
python-dateutil
markdown-it-py
weasyprint